refactor(weather): migrate weather feature to shared module
This commit migrates the weather feature from the application-specific `kocolor` directory to a shared top-level feature module. This refactor centralizes weather-related UI components and logic, allowing for better code reuse across the project.

- **`HomeScreen.kt` & `HomeViewModel.kt`**:
    - Updated imports and component references to point to the new shared package path `com.zoewave.probase.features.weather`.
    - Refactored `LayeredWeatherInfoIcon` and `LayeredWeatherUiState` usage to align with the shared module structure.
- **`build.gradle.kts` (Home Feature)**:
    - Updated project dependencies to point to `:features:weather` instead of the legacy `:applications:kocolor:features:weather` path.
- **`applications/kocolor/features/weather/`**:
    - Deleted the entire local weather feature module, including:
        - Specialized UI components (animations for rain, snow, clouds; wind dials; and various weather card implementations).
        - Feature-specific architecture layers including `WeatherViewModel`, `WeatherUiState`, and `GetCurrentWeatherUseCase`.
        - Module configuration files, resources (strings.xml), and tests.
- **Project Assets**:
    - Removed an unused image asset `img.png` from the `kocolor` directory.
