--- applications/kocolor/features/weather/.gitignore
@@ -1,1 +1,0 @@
-/build
