--- applications/kocolor/features/weather/src/androidTest/java/com/zoewave/probase/kocolor/features/weather/ExampleInstrumentedTest.kt
@@ -1,22 +1,0 @@
-package com.zoewave.probase.kocolor.features.weather
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.zoewave.probase.kocolor.features.weather.test", appContext.packageName)
-    }
-}
