--- applications/kocolor/features/weather/src/main/java/com/zoewave/probase/kocolor/features/weather/ui/WeatherUiState.kt
@@ -1,20 +1,0 @@
-package com.zoewave.probase.kocolor.features.weather.ui
-
-import com.google.android.gms.maps.model.LatLng
-import com.zoewave.probase.core.model.weather.OpenWeatherResponse
-import com.zoewave.probase.core.model.weather.Weather
-
-
-// Define UI states for weather.
-sealed class WeatherUiState {
-    object Loading : WeatherUiState()
-    data class Success(
-        val weatherOpen: OpenWeatherResponse?,
-        var locationString: String = "Santa Barbara, US",
-        val weather: Weather,
-        val settings: Map<String, List<String>>,
-        val location: LatLng? = null
-    ) : WeatherUiState()
-
-    data class Error(val message: String) : WeatherUiState()
-}
