--- applications/kocolor/features/weather/src/main/java/com/zoewave/probase/kocolor/features/weather/ui/components/sun/README.md
@@ -1,24 +1,0 @@
-### Key Improvements
-
-- **Gradient Background:**  
-  A horizontal gradient (from a warm amber to a slightly deeper orange) gives the card a modern,
-  dynamic look.
-
-- **Rounded Corners & Shadows:**  
-  Increased corner rounding to 16.dp along with a more pronounced drop shadow for a softer, elevated
-  feel.
-
-- **Refined Typography:**  
-  The "Temperature" label and value now use bold typography with subtle drop shadows for improved
-  readability against the gradient background.
-
-- **Simplified Sunshine Animation:**  
-  The sunshine decoration has been resized and repositioned for a more elegant accent that doesn’t
-  overwhelm the card’s content.
-
-- **Thermometer Placeholder:**  
-  A simple thermometer placeholder is provided so you can integrate your custom thermometer logic or
-  visuals later.
-
-This version should feel more polished and aligned with Material 3 design principles while providing
-a dynamic, interactive experience for displaying the current temperature.
