--- applications/kocolor/features/weather/src/main/java/com/zoewave/probase/kocolor/features/weather/usecase/GetCurrentWeatherUseCase.kt
@@ -1,31 +1,0 @@
-package com.zoewave.probase.kocolor.features.weather.usecase
-
-
-import com.zoewave.probase.core.network.repository.weather.WeatherRepo
-import javax.inject.Inject
-
-class GetCurrentWeatherUseCase @Inject constructor(
-    private val weatherRepo: WeatherRepo
-) {
-    suspend operator fun invoke(city: String): WeatherInfo {
-        val response = weatherRepo.openCurrentWeatherByCity(city)
-        // Map the API response to our domain model.
-        return WeatherInfo(
-            temperature = response?.main?.temp,
-            condition = response?.weather?.firstOrNull()?.main ?: "Clear",
-            location = "${response?.name}, ${response?.sys?.country}",
-            windDegree = response?.wind?.deg,
-            windSpeed = response?.wind?.speed?.toFloat()
-        )
-    }
-}
-
-data class WeatherInfo(
-    val temperature: Double?,
-    val condition: String?,
-    val location: String?,
-    val windDegree: Int?,
-    val windSpeed: Float?,
-    // Add other properties as needed
-)
-
